:recycle: Move PassphraseInput to toolbox
