+++ src/components/toolbox/passphraseInput/passphraseInput.css
@@ -1,5 +1,5 @@
-@import './../app/variables.css';
-@import './../app/mixins.css';
+@import '../../app/variables.css';
+@import '../../app/mixins.css';
 
 .inputs {
   flex-wrap: wrap;
